Display current news letters in corresponding page
Use local variable in the core controller to get the news letters from
the database.

Clicking on the link will trigger a download of the pdf news letter file.

Currently in Chrome the pdf will open up in the browser and will not be
downloaded to the user specified path.